requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.0
SQLAlchemy==2.0.0
gunicorn==22.0.0
librosa==0.10.0
numpy==1.24.0
pydub==0.25.1
python-dotenv==1.0.0
werkzeug==2.3.0
PyJWT==2.8.0
scikit-learn==1.7.2
joblib==1.5.2
pytest==8.3.5
